Fix cascading deletes when deleting devices
The device inputs and outputs should be deleted if the device is
deleted, so add tests to confirm that.
This test failed initially, and the solution was to add
'passive_deletes=True' to the SQLAlchemy relationship declaration, which
tells SQLAlchemy to trust the database to cascade deletes rather than
trying trying to smooth things over by adding NULLs, which was causing
errors due to violating non-null constraints.