Add ability to run rpi as both scanner and speaker
This uses the selectors Python library, which allows us to wait on
multiple file-like objects (e.g. sockets, serial ports) and handle
incoming data concurrently.

Note: this doesn't do threading or parallel stuff, it's only ever
handling one thing at a time, but that's not a bad thing, as long as the
handlers don't run forever.